srv/database-writer/pyproject.toml
[tool.poetry]
name = "database-writer"
version = "0.1.0"
description = "Database Writer for Website Monitoring"
authors = ["Andrew Yourkin <andrew@yourkin.io>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.75.1"
uvicorn = "^0.17.6"
kafka-python = "^2.0.2"

[tool.poetry.dev-dependencies]
pytest = "^7.1.1"
pytest-mypy = "^0.9.1"
pytest-flake8 = "^1.1.1"
requests = "^2.27.1"
ipython = "^8.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
